requirements.txt
codegen==1.0
coreapi==2.3.0
coreschema==0.0.4
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.10.7
django-allauth==0.32.0
django-angular==0.8.4
django-audiotracks==0.2.4
django-crispy-forms==1.6.1
django-extensions==1.7.9
django-filer==1.2.7
django-filter==1.0.2
django-haystack==2.6.1
django-jenkins==0.19.0
django-markdown-deux==1.0.5
django-mptt==0.8.7
django-multiupload==0.5.2
django-pagedown==0.1.1
django-polymorphic==1.0.2
django-registration-redux==1.6
django-rest-swagger==2.1.2
django-socialprofile==1.2.1
django-tastypie==0.13.3
djangoajax==2.3.7
djangorestframework==3.6.2
djangorestframework-jwt==1.8.0
easy-thumbnails==2.3
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.8
markdown2==2.3.4
MarkupSafe==1.0
mutagen==1.37
mutagenx==1.22
oauthlib==2.0.2
olefile==0.44
openapi-codec==1.3.1
patterns==0.3
pbr==3.1.1
Pillow==4.1.1
PyJWT==1.4.0
pysolr==3.6.0
python-dateutil==2.6.0
python-mimeparse==1.6.0
python-social-auth==0.2.19
python3-openid==3.1.0
pytz==2017.2
requests==2.14.2
requests-oauthlib==0.8.0
simplejson==3.10.0
six==1.10.0
social-auth-core==1.3.0
stevedore==1.23.0
Unidecode==0.4.20
uritemplate==3.0.0
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
virtualenvwrapper-win==1.2.1
whitenoise==3.3.0
